Added Dashing and Particle Effects

+++ Project Thor/Assets/Game/Player/PlayerManager.cs
@@ -243,12 +243,12 @@ public class PlayerManager : NetworkBehaviour
     {
         if (IsOwner)
         {
-            FirstPersonHealthBarText.text = current.ToString() + " / " + MaxHealth.ToString();
+            FirstPersonHealthBarText.text = ((int)(current)).ToString() + " / " + MaxHealth.ToString();
         }
 
         else
         {
-            ThirdPersonHealthBarText.text = current.ToString() + " / " + MaxHealth.ToString();
+            ThirdPersonHealthBarText.text = ((int)(current)).ToString() + " / " + MaxHealth.ToString();
         }
 
         if (previous > 0 && current <= 0)
